Improve location input and map zoom level
Adjusted location input guidance and modified default map zoom and centering. The input validation behavior has been changed so the user-input does not need to start with 'Balai Desa', and instead only the name of the village needs to be entered. Default map details have been updated to have a broader zoom and better coordinates centered on Indonesia.