Implement email check and update mechanism
Updated the code in 'formpendataan.php' and 'pendataan.php' to include checks for existing email entries. If an existing email is uncovered during data input, the related database row will be deleted and replaced with new data. Consequently, when an email that previously exists is being entered, the form submission button will change its text to 'Update' instead of 'Submit'.